==> Empirical Evaluation of Security Analyzers/4_Run_SA_tools.py <==
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
@author: ...

"""

# sudo usermod -a -G docker $USER
# git clone https://github.com/smartbugs/smartbugs
# cd smartbugs
# install/setup-venv.sh
# ./smartbugs -t confuzzius mythril osiris oyente slither -f Solidity/*.sol --processes 12 --mem-limit 96g --timeout 900
# ./reparse results/
# ./results2csv -p results > results.csv


# Analysis completed in 18:51:00.
